~ MinGW compatibility

+++ test/component/test.component.glob/entry.cpp
@@ -4,7 +4,7 @@
  * Purpose: Unit-test of `glob()`.
  *
  * Created: 20th May 2012
- * Updated: 11th May 2025
+ * Updated: 18th May 2025
  *
  * ////////////////////////////////////////////////////////////////////// */
 
@@ -457,7 +457,7 @@ static void TEST_readdir_CWD_FOR_DIRECTORIES_AND_FILES_AND_DOTS()
         std::vector<std::string>        gl_entries;
 
 
-        for (size_t i = 0; gl.gl_pathc != i; ++i)
+        for (int i = 0; gl.gl_pathc != i; ++i)
         {
             // fprintf(stderr, "%d: %s\n", __LINE__, gl.gl_pathv[i]);
 
@@ -514,7 +514,7 @@ static void TEST_readdir_CWD_FOR_DIRECTORIES_AND_DOTS()
         std::vector<std::string>        gl_entries;
 
 
-        for (size_t i = 0; gl.gl_pathc != i; ++i)
+        for (int i = 0; gl.gl_pathc != i; ++i)
         {
             gl_entries.push_back(gl.gl_pathv[i]);
         }
@@ -569,7 +569,7 @@ static void TEST_readdir_HOMEDIR_FOR_DIRECTORIES_AND_FILES_AND_DOTS()
 
         std::vector<std::string>        gl_entries;
 
-        for (size_t i = 0; gl.gl_pathc != i; ++i)
+        for (int i = 0; gl.gl_pathc != i; ++i)
         {
             char const* path = gl.gl_pathv[i];
